The the color type was corrected.

+++ badges/neopixel_matrix_feather/slither/cpp_multi_slither/Worm.h
@@ -28,7 +28,7 @@ class Worm
         
         int displayHeight;
         
-        int segmentColor;
+        uint16_t segmentColor;
 
         uint16_t Color(uint8_t r, uint8_t g, uint8_t b);
